modularizado! un script tiene la funcion y otro la usa
