+++ shuffle.py
@@ -1,29 +1,18 @@
-import random
+import random   # módulo con funciones random
+import mezc     # importamos nuestro módulo
 
-# Si uno empieza una línea con el
-# símbolo # entonces el intérprete
-# ignora esa línea
+CANTIDAD = 8
 
-# A estas líneas se las llama comentarios
-# y sirven para documentar y anotar el código
-
-# Función 'mezclar': recibe una frase en str,
-# y la devuelve con las palabras mezcladas y
-# separadas por guiones, tipo 'str'
-def mezclar(frase):
-    separador = '-'
-    frase_lista = frase.split()
-    random.shuffle(frase_lista)     # esta instrucción la mezcla
-    palabras = separador.join(frase_lista)
-    return palabras
-
+# Si uno empieza una línea con el símbolo # entonces el intérprete (python3) ignora esa línea
+# A estas líneas se las llama comentarios y sirven para documentar y anotar el código
 
 # Pedimos entrada del teclado
-# El valor ingreso termina en la
-# variable 'entrada', y con tipo str
+# El valor ingreso termina en la variable 'entrada', con tipo str
 print("Ingresá una frase")
 print("> ",end='')
 frase_ingresada = input()
 
-frase_mezclada = mezclar(frase_ingresada)
-print(frase_mezclada)
+# Mezclamos las palabras y las volvemos a unir con la función mezclar, que importamos de mezc.py
+for i in range(CANTIDAD):
+    frase_mezclada = mezc.mezclar(frase_ingresada)
+    print(frase_mezclada)
